Update imports to keep Torch working
